Completes happy path for creating new bitmaps and displaying them
The BitmapEditor class is being treated much like a controller, and hence is covered by the integration level tests.
A Renderer class is called to handle the transformation from array to string for the 'S' command.
A CommandProcessor class takes responsibilty for processing the new bitmap command ('I').
At this stage only happy paths are considered, assuming commands are entered in a sensible order and with valid arguments.